requirements.txt
uiautomator2
python-docx
graphviz